Struct secret_toolkit_serialization::Base64Of

pub struct Base64Of<S: Serde, T> {
    pub inner: T,
    // some fields omitted
}

A wrapper that automatically deserializes base64 strings to T using one of the Serde types. Use it as a field of your Handle messages (input and output), for example in the msg field of the Receive interface, to remove the boilerplate of serializing or deserializing the Binary to the relevant type T.

Fields
inner: T